Added description Markdown support. Added better error handling to the SignalGeneratorGUI.py. Changed Pulse Streamer code location.
