+++ SignalGenerator.ui
@@ -88,6 +88,9 @@ p, li { white-space: pre-wrap; }
            <property name="autoFormatting">
             <set>QTextEdit::AutoNone</set>
            </property>
+           <property name="readOnly">
+            <bool>true</bool>
+           </property>
           </widget>
          </item>
          <item>
